email link
Browsing the site as a new customer and when I was on the support page I saw the email didn't allow you to mailto. Added an anchor tag to add that functionality. 

Not sure if y'all did that on purpose so figured I add it in and send it y'all way.